Evaluation

+++ create_confusion_matrix_results.py
@@ -16,7 +16,8 @@ MODELS_DIR = Path('.', 'models')
 SAVE_DIR = Path('.', 'confusion_matrix_results')
 SAVE_DIR.mkdir(parents=True, exist_ok=True)
 
-PATHS_NAMES = [('Standard Model', Path(MODELS_DIR, 'word_char_attention')),
+PATHS_NAMES = [('No Attention Model', Path(MODELS_DIR, 'word_char')),
+               ('Standard Model', Path(MODELS_DIR, 'word_char_attention')),
                ('Code Switching', Path(MODELS_DIR, 'word_char_attention_code_switch')),
                ('Bivalency', Path(MODELS_DIR, 'word_char_attention_bivalency')),
                ('Bivalency and Code Switching', Path(MODELS_DIR, 'word_char_attention_bivalency_code'))]
@@ -27,14 +28,15 @@ for name, model_path in PATHS_NAMES:
     archive = load_archive(model_path.resolve())
     predictor = Predictor.from_archive(archive, 'dialect-predictor')
     save_file = Path(SAVE_DIR, f'{name}.json')
-    with save_file.open('w+') as save_data:
-        for index, data_to_predict in enumerate(TEST_DATA):
-            test_label = data_to_predict['label']
-            prediction = predictor.predict_json({'text': data_to_predict['text']})
-            if index != 0:
-                save_data.write('\n')
-            save_data.write(json.dumps({'prediction': prediction['label'], 
-                                        'label': test_label}))
+    if not save_file.exists():
+        with save_file.open('w+') as save_data:
+            for index, data_to_predict in enumerate(TEST_DATA):
+                test_label = data_to_predict['label']
+                prediction = predictor.predict_json({'text': data_to_predict['text']})
+                if index != 0:
+                    save_data.write('\n')
+                save_data.write(json.dumps({'prediction': prediction['label'], 
+                                            'label': test_label}))
